chore: consolidate evm helpers
