+++ src/components/MultiHopTrade/hooks/useAllowanceApproval/hooks/useApprovalTx.tsx
@@ -5,7 +5,7 @@ import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingl
 import { usePoll } from 'hooks/usePoll/usePoll'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import type { TradeQuote } from 'lib/swapper/api'
-import { isEvmChainAdapter } from 'lib/utils'
+import { isEvmChainAdapter } from 'lib/utils/evm'
 
 import { APPROVAL_POLL_INTERVAL_MILLISECONDS } from '../../constants'
 import { getApprovalTxData } from '../helpers'
