+++ src/components/MultiHopTrade/hooks/useAllowanceApproval/hooks/useExecuteAllowanceApproval.tsx
@@ -6,8 +6,7 @@ import { useErrorHandler } from 'hooks/useErrorToast/useErrorToast'
 import { usePoll } from 'hooks/usePoll/usePoll'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import type { TradeQuote } from 'lib/swapper/api'
-import { isEvmChainAdapter } from 'lib/utils'
-import { buildAndBroadcast } from 'lib/utils/evm'
+import { buildAndBroadcast, isEvmChainAdapter } from 'lib/utils/evm'
 
 import { APPROVAL_POLL_INTERVAL_MILLISECONDS } from '../../constants'
 import { checkApprovalNeeded } from '../helpers'
