+++ src/lib/utils/index.ts
@@ -1,8 +1,6 @@
 import { skipToken } from '@reduxjs/toolkit/dist/query'
 import type { AssetReference } from '@shapeshiftoss/caip'
 import { ASSET_REFERENCE } from '@shapeshiftoss/caip'
-import type { EvmChainAdapter, EvmChainId } from '@shapeshiftoss/chain-adapters'
-import { evmChainIds } from '@shapeshiftoss/chain-adapters'
 import type { HDWallet } from '@shapeshiftoss/hdwallet-core'
 import { KeepKeyHDWallet } from '@shapeshiftoss/hdwallet-keepkey'
 import { KeplrHDWallet } from '@shapeshiftoss/hdwallet-keplr'
@@ -15,8 +13,6 @@ import difference from 'lodash/difference'
 import intersection from 'lodash/intersection'
 import isUndefined from 'lodash/isUndefined'
 import union from 'lodash/union'
-
-export * as evm from './evm'
 
 // we don't want utils to mutate by default, so spreading here is ok
 export const upsertArray = <T extends unknown>(arr: T[], item: T): T[] =>
@@ -161,10 +157,6 @@ export const hashCode = (str: string): string =>
 
 export const sha256 = (input: string): string =>
   crypto.createHash('sha256').update(input).digest('hex')
-
-export const isEvmChainAdapter = (chainAdapter: unknown): chainAdapter is EvmChainAdapter => {
-  return evmChainIds.includes((chainAdapter as EvmChainAdapter).getChainId() as EvmChainId)
-}
 
 // https://github.com/sniptt-official/monads/issues/111
 export const AsyncResultOf = async <T>(promise: Promise<T>): Promise<Result<T, Error>> => {
